Split up templateLinter.test.ts
It was getting too big for one file.
